+++ dj-db/src/main/java/cn/dlbdata/dj/db/pojo/DjActiveUser.java
@@ -9,19 +9,19 @@ public class DjActiveUser {
      * 记录ID
      */
     @Id
-    private Integer id;
+    private Long id;
 
     /**
      * 活动ID
      */
     @Column(name = "dj_active_id")
-    private Integer djActiveId;
+    private Long djActiveId;
 
     /**
      * 参与人ID
      */
     @Column(name = "dj_user_id")
-    private Integer djUserId;
+    private Long djUserId;
 
     /**
      * 状态
@@ -45,7 +45,7 @@ public class DjActiveUser {
      *
      * @return id - 记录ID
      */
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
@@ -54,7 +54,7 @@ public class DjActiveUser {
      *
      * @param id 记录ID
      */
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
 
@@ -63,7 +63,7 @@ public class DjActiveUser {
      *
      * @return dj_active_id - 活动ID
      */
-    public Integer getDjActiveId() {
+    public Long getDjActiveId() {
         return djActiveId;
     }
 
@@ -72,7 +72,7 @@ public class DjActiveUser {
      *
      * @param djActiveId 活动ID
      */
-    public void setDjActiveId(Integer djActiveId) {
+    public void setDjActiveId(Long djActiveId) {
         this.djActiveId = djActiveId;
     }
 
@@ -81,7 +81,7 @@ public class DjActiveUser {
      *
      * @return dj_user_id - 参与人ID
      */
-    public Integer getDjUserId() {
+    public Long getDjUserId() {
         return djUserId;
     }
 
@@ -90,7 +90,7 @@ public class DjActiveUser {
      *
      * @param djUserId 参与人ID
      */
-    public void setDjUserId(Integer djUserId) {
+    public void setDjUserId(Long djUserId) {
         this.djUserId = djUserId;
     }
